initial implementation of watching files outside the user's emails directory
